Added Windows Essentials Training Materials (WATCH OUT)
The slides (pptx) are corret with this commit but the handouts per section (docx) and for the entire course (pdf) are completely wrong. I could not update or generate the documents from a Mac using the latest version of Office. Pushing this commit and then addressing this on a Windows machine with Office.