Updated README - better layout and known issues
Included the abstract that we offer on the trianing page.
Re-ordered the content to make it clearer to describe what the repository contains
Expanded the publishing content section to include more details about the various piece.
Removed the links to the Chef Essentials Linux content
Added a known issues section to that describes the three major issues affecting the training at this moment